Update event_categories.js
Added functionality to apply a categories filter to the calendar view. 
Note there is a bug because you first have to switch to another category before you can filter the first category. 
Also the code does only seem to work with .change() as I tested it many times with .select() but without success.